requirements.txt
requests
python-telegram-bot
PyYAML
google-generativeai
asyncio
